Identity & external login SSO

+++ src/AuctionService/Controllers/AuctionsController.cs
@@ -1,12 +1,11 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using AuctionService.Data;
 using AuctionService.DTOs;
 using AuctionService.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -19,8 +18,10 @@ namespace AuctionService.Controllers
     {
         private readonly AuctionDbContext _context;
         private readonly IMapper _mapper;
-        public AuctionsController(AuctionDbContext context, IMapper mapper)
+        private readonly IPublishEndpoint _publishEndpoint;
+        public AuctionsController(AuctionDbContext context, IMapper mapper, IPublishEndpoint publishEndpoint)
         {
+            _publishEndpoint = publishEndpoint;
             _mapper = mapper;
             _context = context;
 
@@ -53,19 +54,30 @@ namespace AuctionService.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost] 
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto) {
             var auction = _mapper.Map<Auction>(auctionDto);
-            auction.Seller = "test";
+            auction.Seller = User.Identity.Name;
+
+            // transaction begins
             _context.Auctions.Add(auction);
+
+            var newAuction = _mapper.Map<AuctionDto>(auction);
+            // FOR CONSUMERS
+            await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
+
             var result = await _context.SaveChangesAsync() > 0;
+
+
             if(!result) {
                 return BadRequest("Could not save changes to db");
             }
-            return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, _mapper.Map<AuctionDto>(auction));
+
+            return CreatedAtAction(nameof(GetAuctionById), new {auction.Id}, newAuction);
         }
 
-
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id,UpdateAuctionDto updateAuctionDto) {
 
@@ -77,12 +89,15 @@ namespace AuctionService.Controllers
          
 
             //CHECK SELLER == USER
+
+            if(auction.Seller != User.Identity.Name) return Forbid();
+
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
             auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
             auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
-            
+            await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
             var result = await _context.SaveChangesAsync() > 0;
 
             if(result) return Ok();
@@ -90,15 +105,17 @@ namespace AuctionService.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
 
         public async Task<ActionResult> DeleteAuction(Guid id) {
             //to do implement identity to chekc if deleter is the owner
             var auction = await _context.Auctions.FindAsync(id);
             if (auction == null) return NotFound();
-
+            
+            if(auction.Seller != User.Identity.Name) return Forbid();
             _context.Auctions.Remove(auction);
-
+            await _publishEndpoint.Publish<AuctionDeleted>(new {Id = auction.Id.ToString()});
             var result = await _context.SaveChangesAsync() > 0;
             if(result) return Ok();
